Preset name filtering and upsert method

+++ golem/config/presets.py
@@ -1,5 +1,6 @@
 import logging
 
+from golem.appconfig import CUSTOM_HARDWARE_PRESET_NAME, DEFAULT_HARDWARE_PRESET_NAME
 from golem.core.hardware import HardwarePresets
 from golem.model import HardwarePreset
 
@@ -27,13 +28,41 @@ class HardwarePresetsMixin(object):
         preset.update()
         return preset
 
-    def update_hw_preset(self, name, preset_dict):
+    @classmethod
+    def update_hw_preset(cls, preset):
+        preset_dict = cls.__preset_to_dict(preset)
+        name = cls.__sanitize_preset_name(preset_dict['name'])
+
         preset = HardwarePreset.get(name=name)
         preset.apply(preset_dict)
         preset.update()
+        return preset
 
-    def remove_hw_preset(self, name):
+    @classmethod
+    def upsert_hw_preset(cls, preset):
+        preset_dict = cls.__preset_to_dict(preset)
+        name = cls.__sanitize_preset_name(preset_dict['name'])
+
+        preset, _ = HardwarePreset.get_or_create(name=name)
+        preset.apply(preset_dict)
+        preset.update()
+        return preset
+
+    @staticmethod
+    def remove_hw_preset(name):
+        if name in [CUSTOM_HARDWARE_PRESET_NAME, DEFAULT_HARDWARE_PRESET_NAME]:
+            raise ValueError('Cannot remove preset with name: ' + name)
         HardwarePreset.delete().where(name=name)
 
-    def activate_hw_preset(self, name):
+    def activate_hw_preset(self, name, run_benchmarks=False):
         raise NotImplementedError
+
+    @staticmethod
+    def __preset_to_dict(preset):
+        return preset if isinstance(preset, dict) else preset.to_dict()
+
+    @staticmethod
+    def __sanitize_preset_name(name):
+        if not name or name == DEFAULT_HARDWARE_PRESET_NAME:
+            return CUSTOM_HARDWARE_PRESET_NAME
+        return name
